+++ internal/grpcwrapper/rawtopic/create_topic.go
@@ -6,6 +6,7 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Topic"
 	"google.golang.org/protobuf/types/known/durationpb"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawoptional"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawtopic/rawtopiccommon"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawydb"
 )
@@ -23,6 +24,7 @@ type CreateTopicRequest struct {
 	Attributes                        map[string]string
 	Consumers                         []Consumer
 	MeteringMode                      MeteringMode
+	MetricsLevel                      rawoptional.Uint32
 }
 
 func (req *CreateTopicRequest) ToProto() *Ydb_Topic.CreateTopicRequest {
@@ -46,6 +48,8 @@ func (req *CreateTopicRequest) ToProto() *Ydb_Topic.CreateTopicRequest {
 	}
 
 	proto.MeteringMode = Ydb_Topic.MeteringMode(req.MeteringMode)
+
+	proto.MetricsLevel = req.MetricsLevel.ToProto()
 
 	return proto
 }